+++ src/Exceptions/HydrationException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Swis\JsonApi\Client\Exceptions;
+
+use InvalidArgumentException;
+
+class HydrationException extends InvalidArgumentException implements Exception
+{
+}