+++ src/Exceptions/TypeMappingException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Swis\JsonApi\Client\Exceptions;
+
+use InvalidArgumentException;
+
+class TypeMappingException extends InvalidArgumentException implements Exception
+{
+}