+++ src/Exceptions/UnsupportedDataException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Swis\JsonApi\Client\Exceptions;
+
+use InvalidArgumentException;
+
+class UnsupportedDataException extends InvalidArgumentException implements Exception
+{
+}